==> 1_ultimate_guitar_scrape.ipynb <==
# Ultimate Guitar Scrape - 1
## Scraping all time tabs
*Initial scrape of top 5,000 tabs and links based on views on ultimate-guitar.com*

----

**Final Project for Data & Databases**

**C.J. Robinson**

**Fall 2024**

-----
### Scrape main page

import re
import requests
from bs4 import BeautifulSoup
import time
import pandas as pd

Page relies heavily on JS fter loading

### Shifting to Playwright

from playwright.async_api import async_playwright

playwright = await async_playwright().start()
browser = await playwright.chromium.launch(headless=False)
page = await browser.new_page()
await page.goto("https://www.ultimate-guitar.com/explore?order=hitstotal_desc")

html = await page.content()
soup_doc = BeautifulSoup(html, "html.parser")

table = soup_doc.find_all('article')[1]

all_songs = []

rank_counter = 1
for row in table.find_all("div", class_ = "LQUZJ")[1:]:
    song_dict = {}

    # there's no rank, so imputing one here
    song_dict['rank'] = rank_counter

    # based off of class names which...seem static
    song_dict['artist'] = row.find(class_ ="lIKMM lz4gy").text
    song_dict['song'] = row.find(class_ ="lIKMM g7KZB").text
    song_dict['ratings'] = row.find(class_ ="lIKMM eznJV").text
    song_dict['hits'] = row.find(class_ ="lIKMM UpKH8").text.strip()
    song_dict['type'] = row.find(class_ ="lIKMM PdXKy").text

    song_dict['song_link'] = row.find("a", class_ = "aPPf7 HT3w5 lBssT")['href']
    song_dict['artist_link'] = row.find("a", class_ = "aPPf7 jtEAE lBssT")['href']

    #stars are a bit interesting
    # take all full stars
    star_count = len(row.find_all("span", class_ = "kd3Q7 DSnE7")) * 1.0
    # add half stars
    star_count += len(row.find_all("span", class_ = "kd3Q7 RCXwf DSnE7")) * .5
    # we wont care about empty stars - assign star
    song_dict['star_count'] = star_count
    
    #up the counter
    rank_counter += 1

    all_songs.append(song_dict)

print(all_songs)

pd.json_normalize(all_songs)

### Loop through all the pages!

# start up the page!
playwright = await async_playwright().start()
browser = await playwright.chromium.launch(headless=False)
page = await browser.new_page()

all_songs = []
rank_counter = 1

for page_num in range(1,101):
    url = f"https://www.ultimate-guitar.com/explore?order=hitstotal_desc&page={str(page_num)}"
    print(url)
    await page.goto(url)
    await page.get_by_text("Artist", exact=True).wait_for()
    html = await page.content()
    soup_doc = BeautifulSoup(html, "html.parser")
    table = soup_doc.find_all('article')[1]
    
    for row in table.find_all("div", class_ = "LQUZJ")[1:]:
        song_dict = {}

        # there's no rank, so imputing one here
        song_dict['rank'] = rank_counter
    
        # based off of class names which...seem static
        try:
            #get just raw artist text (feat, &, commas)
            artist_list = []
            song_dict['artist'] = row.find(class_ ="lIKMM lz4gy").text

            # get each individual artist in a list...just in case!
            for artist in row.find_all('a', class_ ="aPPf7 jtEAE lBssT"):
                artist_list.append(artist.text)
            
            song_dict['artist_list'] = artist_list
        except:
            print(f"No artist for {rank_counter} rank")
            
        try:    
            song_dict['song'] = row.find(class_ ="lIKMM g7KZB").text
        except:
            print(f"No song for {rank_counter} rank")

        try:
            song_dict['ratings'] = row.find(class_ ="lIKMM eznJV").text
        except:
            print(f"No ratings for {rank_counter} rank")
            
        try:   
            song_dict['hits'] = row.find(class_ ="lIKMM UpKH8").text.strip()
        except:
            print(f"No hits for {rank_counter} rank")
            
        try:    
            song_dict['type'] = row.find(class_ ="lIKMM PdXKy").text
        except:
            print(f"No type for {rank_counter} rank")
    
        song_dict['song_link'] = row.find("a", class_ = "aPPf7 HT3w5 lBssT")['href']
        song_dict['artist_link'] = row.find("a", class_ = "aPPf7 jtEAE lBssT")['href']
    
        #stars are a bit interesting
        # take all full stars
        star_count = len(row.find_all("span", class_ = "kd3Q7 DSnE7")) * 1.0
        # add half stars
        star_count += len(row.find_all("span", class_ = "kd3Q7 RCXwf DSnE7")) * .5
        # we wont care about empty stars - assign star
        try:
            song_dict['star_count'] = star_count
        except:
            print(f"No stars for {rank_counter} rank")
        
        #up the counter
        rank_counter += 1
    
        all_songs.append(song_dict)

    # wait a couple seconds
    time.sleep(5)

df = pd.json_normalize(all_songs)
df.to_csv("top_songs.csv")
