==> 2_ug_ind_songs.ipynb <==
# Ultimate Guitar Scrape - 2
## Scraping individual tabs
*After initial scrape of the top 5,000 tabs on ultimate-guitar.com, scrape each individual tab for metadata/chords*

----

**Final Project for Data & Databases**

**C.J. Robinson**

**Fall 2024**

-----
### Scrape song pages

import re
import requests
from bs4 import BeautifulSoup
import time
import pandas as pd
from playwright.async_api import async_playwright

# read in initial data
top_songs_df = pd.read_csv("top_songs.csv")
top_songs_df.head()

### Loop through all the pages!

playwright = await async_playwright().start()
browser = await playwright.chromium.launch(headless=False)
page = await browser.new_page()

song_list = []
bad_urls_1k = []
counter = 0

for url in top_songs_df['song_link'][0:1000]:
    my_url = url
    #counter for progress
    counter += 1
    
    try:
        await page.goto(my_url, timeout=120000)
        time.sleep(3)
        #scroll down to get chords to actually load
        await page.evaluate("window.scrollTo(0, 2000)") 
        time.sleep(3)
        html = await page.content()
        
        soup_doc = BeautifulSoup(html, "html.parser")
        
        song_dict = {}
        song_dict['link'] = my_url
        
        #there's a bunch of metadata that may or may not be in the header
        #since it depends on the author
        #go through anything in that header and pull out metadata name and value
        try:
            for tag in soup_doc.find_all("th", class_ = "ZvOWv"):
                data_type = tag.text.lower()
                song_dict[data_type] = tag.next_sibling()[0].text
        except:
            print("No meta data")
            print(my_url)
        
        # only grab chord elements if it is chords, not tabs
        try:
            chord_list = []
            #pull all chords
            for chords in soup_doc.find_all("span", class_="_Oy28"):
                chord_list.append(chords.text)
        
            #get unique chord list
            song_dict['chord_list'] = list(set(chord_list))
            #also grab full text...just in case
            song_dict['full_text'] = soup_doc.find("pre", class_="tK8GG Ty_RP").text
        except:
            print("No chords")
        
        # get raw text of contributions, will regex later
        try:
            song_dict['contributors'] = soup_doc.find("span", class_="zku_4").text
        except:
            print("No contributors" + my_url)
        
        # get popularity raw, also will regex later
        try:
            song_dict['popularity'] = soup_doc.find("div", class_="_apVM").text
        except:
            print("No popularity" + my_url)
    
        song_list.append(song_dict)

    except:
        print("error with " + url)
        # get a list of errored URLS
        bad_urls_1k.append(url)

    if counter % 10 == 0:
        print(str(counter) + " / 1000 done") 
    
top_songs_df_meta = pd.json_normalize(song_list)
top_songs_df_meta.to_csv("top_songs_metadata_1k.csv", index = False)

bad_urls_1k_df = pd.DataFrame(bad_urls_1k, columns=["bad_urls"])
bad_urls_1k_df.to_csv('bad_urls_1k_df.csv', index=False)

# top_songs_df = pd.json_normalize(song_list)
# top_songs_df.to_csv("top_songs_metadata_1k.csv", index = False)

playwright = await async_playwright().start()
browser = await playwright.chromium.launch(headless=False)
page = await browser.new_page()

song_list_2k = []
bad_urls_2k = []
counter = 0

for url in top_songs_df['song_link'][1000:2000]:
    my_url = url
    #counter for progress
    counter += 1
    
    try:
        await page.goto(my_url, timeout=120000)
        time.sleep(3)
        #scroll down to get chords to actually load
        await page.evaluate("window.scrollTo(0, 2000)") 
        time.sleep(3)
        html = await page.content()
        
        soup_doc = BeautifulSoup(html, "html.parser")
        
        song_dict = {}
        song_dict['link'] = my_url
        
        #there's a bunch of metadata that may or may not be in the header
        #since it depends on the author
        #go through anything in that header and pull out metadata name and value
        try:
            for tag in soup_doc.find_all("th", class_ = "ZvOWv"):
                data_type = tag.text.lower()
                song_dict[data_type] = tag.next_sibling()[0].text
        except:
            print("No meta data")
            print(my_url)
        
        # only grab chord elements if it is chords, not tabs
        try:
            chord_list = []
            #pull all chords
            for chords in soup_doc.find_all("span", class_="_Oy28"):
                chord_list.append(chords.text)
        
            #get unique chord list
            song_dict['chord_list'] = list(set(chord_list))
            #also grab full text...just in case
            song_dict['full_text'] = soup_doc.find("pre", class_="tK8GG Ty_RP").text
        except:
            print("No chords")
        
        # get raw text of contributions, will regex later
        try:
            song_dict['contributors'] = soup_doc.find("span", class_="zku_4").text
        except:
            print("No contributors" + my_url)
        
        # get popularity raw, also will regex later
        try:
            song_dict['popularity'] = soup_doc.find("div", class_="_apVM").text
        except:
            print("No popularity" + my_url)
    
        song_list_2k.append(song_dict)

    except:
        print("error with " + url)
        # get a list of errored URLS
        bad_urls_2k.append(url)

    if counter % 10 == 0:
        print(str(counter) + " / 1000 done") 
    
top_songs_df_meta_2k = pd.json_normalize(song_list_2k)
top_songs_df_meta_2k.to_csv("top_songs_metadata_2k.csv", index = False)

bad_urls_2k_df = pd.DataFrame(bad_urls_2k, columns=["bad_urls"])
bad_urls_2k_df.to_csv('bad_urls_2k_df.csv', index=False)

playwright = await async_playwright().start()
browser = await playwright.chromium.launch(headless=False)
page = await browser.new_page()

song_list_3k = []
bad_urls_3k = []
counter = 0

for url in top_songs_df['song_link'][2000:3000]:
    my_url = url
    #counter for progress
    counter += 1
    
    try:
        await page.goto(my_url, timeout=120000)
        time.sleep(2)
        #scroll down to get chords to actually load
        await page.evaluate("window.scrollTo(0, 2000)") 
        time.sleep(2)
        html = await page.content()
        
        soup_doc = BeautifulSoup(html, "html.parser")
        
        song_dict = {}
        song_dict['link'] = my_url
        
        #there's a bunch of metadata that may or may not be in the header
        #since it depends on the author
        #go through anything in that header and pull out metadata name and value
        try:
            for tag in soup_doc.find_all("th", class_ = "ZvOWv"):
                data_type = tag.text.lower()
                song_dict[data_type] = tag.next_sibling()[0].text
        except:
            print("No meta data")
            print(my_url)
        
        # only grab chord elements if it is chords, not tabs
        try:
            chord_list = []
            #pull all chords
            for chords in soup_doc.find_all("span", class_="_Oy28"):
                chord_list.append(chords.text)
        
            #get unique chord list
            song_dict['chord_list'] = list(set(chord_list))
            #also grab full text...just in case
            song_dict['full_text'] = soup_doc.find("pre", class_="tK8GG Ty_RP").text
        except:
            print("No chords")
        
        # get raw text of contributions, will regex later
        try:
            song_dict['contributors'] = soup_doc.find("span", class_="zku_4").text
        except:
            print("No contributors" + my_url)
        
        # get popularity raw, also will regex later
        try:
            song_dict['popularity'] = soup_doc.find("div", class_="_apVM").text
        except:
            print("No popularity" + my_url)
    
        song_list_3k.append(song_dict)

    except:
        print("error with " + url)
        # get a list of errored URLS
        bad_urls_3k.append(url)

    if counter % 10 == 0:
        print(str(counter) + " / 1000 done") 
    
top_songs_df_meta_3k = pd.json_normalize(song_list_3k)
top_songs_df_meta_3k.to_csv("top_songs_metadata_3k.csv", index = False)

bad_urls_3k_df = pd.DataFrame(bad_urls_3k, columns=["bad_urls"])
bad_urls_3k_df.to_csv('bad_urls_3k_df.csv', index=False)

playwright = await async_playwright().start()
browser = await playwright.chromium.launch(headless=False)
page = await browser.new_page()

song_list_4k = []
bad_urls_4k = []
counter = 0

for url in top_songs_df['song_link'][3000:5000]:
    my_url = url
    #counter for progress
    counter += 1
    
    try:
        await page.goto(my_url, timeout=120000)
        time.sleep(2)
        #scroll down to get chords to actually load
        await page.evaluate("window.scrollTo(0, 2000)") 
        time.sleep(2)
        html = await page.content()
        
        soup_doc = BeautifulSoup(html, "html.parser")
        
        song_dict = {}
        song_dict['link'] = my_url
        
        #there's a bunch of metadata that may or may not be in the header
        #since it depends on the author
        #go through anything in that header and pull out metadata name and value
        try:
            for tag in soup_doc.find_all("th", class_ = "ZvOWv"):
                data_type = tag.text.lower()
                song_dict[data_type] = tag.next_sibling()[0].text
        except:
            print("No meta data")
            print(my_url)
        
        # only grab chord elements if it is chords, not tabs
        try:
            chord_list = []
            #pull all chords
            for chords in soup_doc.find_all("span", class_="_Oy28"):
                chord_list.append(chords.text)
        
            #get unique chord list
            song_dict['chord_list'] = list(set(chord_list))
            #also grab full text...just in case
            song_dict['full_text'] = soup_doc.find("pre", class_="tK8GG Ty_RP").text
        except:
            print("No chords")
        
        # get raw text of contributions, will regex later
        try:
            song_dict['contributors'] = soup_doc.find("span", class_="zku_4").text
        except:
            print("No contributors" + my_url)
        
        # get popularity raw, also will regex later
        try:
            song_dict['popularity'] = soup_doc.find("div", class_="_apVM").text
        except:
            print("No popularity" + my_url)
    
        song_list_4k.append(song_dict)

    except:
        print("error with " + url)
        # get a list of errored URLS
        bad_urls_4k.append(url)

    if counter % 10 == 0:
        print(str(counter) + " done") 
    
top_songs_df_meta_4k = pd.json_normalize(song_list_4k)
top_songs_df_meta_4k.to_csv("top_songs_metadata_4k.csv", index = False)

bad_urls_4k_df = pd.DataFrame(bad_urls_4k, columns=["bad_urls"])
bad_urls_4k_df.to_csv('bad_urls_4k_df.csv', index=False)
