==> 3_mb_artist_data.ipynb <==
# Ultimate Guitar Scrape - 3
## Scrape music 
*Take artist names and search for them on the Music Brainz API, an open source encyclopedia for artist information*

----

**Final Project for Data & Databases**

**C.J. Robinson**

**Fall 2024**

import re
import requests
from bs4 import BeautifulSoup
import time
import pandas as pd
from playwright.async_api import async_playwright
import numpy as np

df = pd.read_csv("top_songs.csv")

#take all unique artists so i dont have to do this multiple times
artists = list(set(df['artist']))

all_artists = []
counter = 0
for artist in artists:
    artist_dict = {}
    artist_dict['artist'] = artist
    url = "https://musicbrainz.org/ws/2/artist?fmt=json&query=" + artist.replace(" ","+") 
    response = requests.get(url)
    data = response.json()
    try:
        artist_dict['id'] = data['artists'][0]['id']
    except:
        print("No id for " + artist)

    try:
        artist_dict['type'] = data['artists'][0]['type']
    except:
        print("No type for " + artist)

    try:
        artist_dict['gender'] = data['artists'][0]['gender']
    except:
        print("No gender for " + artist)    

    try:
        artist_dict['country'] = data['artists'][0]['country']
    except:
        print("No country for " + artist)  

    try:
        artist_dict['start_date'] = data['artists'][0]['life-span']['begin']
    except:
        print("No start date for " + artist)  
        
    try:
        tag_list = []
        for tag in  data['artists'][0]['tags']:
            tag_list.append(tag['name'])   
        artist_dict['tags'] = tag_list
    except:
        print("No tags for " + artist)  
        

    all_artists.append(artist_dict)
    time.sleep(1)

    counter+=1
    if counter % 100 == 0:
        print(str(counter) + " done!")
        artist_data = pd.json_normalize(all_artists)
        artist_data.to_csv("artist_data.csv", index = False)


artist_data = pd.json_normalize(all_artists)
artist_data.to_csv("artist_data.csv", index = False)
