==> 4_collate.ipynb <==
# Ultimate Guitar Scrape - 4
## Scrape any missed tabs and collate
*Take all tabs missed from individual scrape and rescrape, then collate all CSVs together*

----

**Final Project for Data & Databases**

**C.J. Robinson**

**Fall 2024**

-----
### Scrape song pages

import re
import requests
from bs4 import BeautifulSoup
import time
import pandas as pd
from playwright.async_api import async_playwright
import numpy as np

# read in all bad urls
bad_urls = pd.DataFrame()

bad_urls = pd.concat([pd.read_csv("bad_urls_1k_df.csv"),
                     pd.read_csv("bad_urls_2k_df.csv"),
                     pd.read_csv("bad_urls_3k_df.csv"),
                     pd.read_csv("bad_urls_4k_df.csv")], ignore_index=True)

### Loop through all URLs missed

playwright = await async_playwright().start()
browser = await playwright.chromium.launch(headless=False)
page = await browser.new_page()

retry_song_list = []

for url in bad_urls['bad_urls']:
    my_url = url
    
    try:
        await page.goto(my_url, timeout=120000)
        time.sleep(3)
        #scroll down to get chords to actually load
        await page.evaluate("window.scrollTo(0, 2000)") 
        time.sleep(3)
        html = await page.content()
        
        soup_doc = BeautifulSoup(html, "html.parser")
        
        song_dict = {}
        song_dict['link'] = my_url
        
        #there's a bunch of metadata that may or may not be in the header
        #since it depends on the author
        #go through anything in that header and pull out metadata name and value
        try:
            for tag in soup_doc.find_all("th", class_ = "ZvOWv"):
                data_type = tag.text.lower()
                song_dict[data_type] = tag.next_sibling()[0].text
        except:
            print("No meta data")
            print(my_url)
        
        # only grab chord elements if it is chords, not tabs
        try:
            chord_list = []
            #pull all chords
            for chords in soup_doc.find_all("span", class_="_Oy28"):
                chord_list.append(chords.text)
        
            #get unique chord list
            song_dict['chord_list'] = list(set(chord_list))
            #also grab full text...just in case
            song_dict['full_text'] = soup_doc.find("pre", class_="tK8GG Ty_RP").text
        except:
            print("No chords")
        
        # get raw text of contributions, will regex later
        try:
            song_dict['contributors'] = soup_doc.find("span", class_="zku_4").text
        except:
            print("No contributors" + my_url)
        
        # get popularity raw, also will regex later
        try:
            song_dict['popularity'] = soup_doc.find("div", class_="_apVM").text
        except:
            print("No popularity" + my_url)
    
        retry_song_list.append(song_dict)

    except:
        print("error with " + url)
    
retry_song_list = pd.json_normalize(retry_song_list)
retry_song_list.to_csv("retry_song_list.csv", index = False)

### Read in all CSVs

top_songs = pd.read_csv("top_songs.csv")
#drop the index column
top_songs = top_songs.drop(top_songs.columns[0], axis=1) 

song_metadata = pd.concat([pd.read_csv("retry_song_list.csv"),
                       pd.read_csv("top_songs_metadata_1k.csv"),
                       pd.read_csv("top_songs_metadata_2k.csv"),
                       pd.read_csv("top_songs_metadata_3k.csv"),
                       pd.read_csv("top_songs_metadata_4k.csv")])

artist_data = pd.read_csv('artist_data.csv')
artist_data = artist_data.rename(columns={'type':'person_group'})

### Join on original

df = top_songs.merge(song_metadata, how = "left", left_on = "song_link", right_on = "link") 
df = df.merge(artist_data, how = "left", left_on = "artist", right_on = "artist")

# clean up some columns
df['contributors'] = df['contributors'].str.extract(r'(\d+)')
df['favorites'] = df['popularity'].str.extract(r'\d[\d,]* views, added to favorites (\d[\d,]*) times')
df['favorites'] = df['favorites'].str.replace(',', '')

# change to int
df['ratings'] = df['ratings'].str.replace(',', '')
df['hits'] = df['hits'].str.replace(',', '')

# get song title wihout version
df['song_no_vers'] = df['song'].str.replace(r'\s*\(ver\s*\d+\)\s*', '', regex=True)

#get version (if no version, it's the first version)
df['version'] = df['song'].str.extract(r'ver\s*(\d+)', expand=False).fillna(1)

#change capo to int
df['capo'] = df['capo: '].str.extract(r'(^\d+)')

# if there are 0 stars, it's just that there's no ratings
df['star_count'] = df['star_count'].replace(0, np.nan)

df.rename(columns= {
    'key: ':'key', 
    'tuning: ':'tuning',
    'difficulty: ':'difficulty'}, inplace=True)

# get rid of some unnecessary columns
df = df.drop(columns=['file format: ', 'filesize: ', 'instuments: ', 'capo: '], axis=1) 


df.head(10)

df.to_csv("top_songs_full.csv", index=False)
